requirements.txt
dataclasses_json==0.6.7
discord_py_interactions==5.12.1
Requests==2.32.3
python-dotenv>=0.19.2